Move the test code to ParquetFileFormatSuite.
