+++ sql/core/src/test/scala/org/apache/spark/sql/execution/datasources/parquet/ParquetIOSuite.scala
@@ -17,7 +17,7 @@
 
 package org.apache.spark.sql.execution.datasources.parquet
 
-import java.time.{Duration, LocalDateTime, Period}
+import java.time.LocalDateTime
 import java.util.Locale
 
 import scala.collection.JavaConverters._
@@ -1064,29 +1064,6 @@ class ParquetIOSuite extends QueryTest with ParquetTest with SharedSparkSession 
           assert(Seq(866, 20, 492, 76, 824, 604, 343, 820, 864, 243)
             .zip(last10Df).forall(d =>
             d._1 == d._2.getDecimal(0).unscaledValue().intValue()))
-      }
-    }
-  }
-
-  test("SPARK-36825, SPARK-36854: year-month/day-time intervals written and read as INT32/INT64") {
-    Seq(false, true).foreach { offHeapEnabled =>
-      withSQLConf(SQLConf.COLUMN_VECTOR_OFFHEAP_ENABLED.key -> offHeapEnabled.toString) {
-        Seq(
-          YearMonthIntervalType() -> ((i: Int) => Period.of(i, i, 0)),
-          DayTimeIntervalType() -> ((i: Int) => Duration.ofDays(i).plusSeconds(i))
-        ).foreach { case (it, f) =>
-          val data = (1 to 10).map(i => Row(i, f(i)))
-          val schema = StructType(Array(StructField("d", IntegerType, false),
-            StructField("i", it, false)))
-          withTempPath { file =>
-            val df = spark.createDataFrame(sparkContext.parallelize(data), schema)
-            df.write.parquet(file.getCanonicalPath)
-            withAllParquetReaders {
-              val df2 = spark.read.parquet(file.getCanonicalPath)
-              checkAnswer(df2, df.collect().toSeq)
-            }
-          }
-        }
       }
     }
   }
